+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'chirpsdk'
+rootProject.name = 'chirp_flutter'
