+++ example/.metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 5391447fae6209bb21a89e6a5a6583cac1af9b4b
+  revision: 1946fc4da0f80c522d7e3ae7d4f7309908ed86f2
   channel: dev
 
 project_type: app
