+++ example/README.md
@@ -1,9 +1,16 @@
-# ChirpSDK Example
+# chirp_flutter_example
 
-Demonstrates how to use the ChirpSDK plugin.
+Demonstrates how to use the chirp_flutter plugin.
 
 ## Getting Started
 
-This project is a starting point for a Flutter application using Chirp.
+This project is a starting point for a Flutter application.
 
-- [Example Code](https://github.com/chirp/chirp-flutter/tree/master/example)
+A few resources to get you started if this is your first Flutter project:
+
+- [Lab: Write your first Flutter app](https://flutter.dev/docs/get-started/codelab)
+- [Cookbook: Useful Flutter samples](https://flutter.dev/docs/cookbook)
+
+For help getting started with Flutter, view our
+[online documentation](https://flutter.dev/docs), which offers tutorials,
+samples, guidance on mobile development, and a full API reference.
