+++ example/android/app/src/main/kotlin/com/example/chirp_flutter_example/MainActivity.kt
@@ -1,7 +1,6 @@
-package io.chirp.chirpsdkexample
+package com.example.chirp_flutter_example
 
 import android.os.Bundle
-
 import io.flutter.app.FlutterActivity
 import io.flutter.plugins.GeneratedPluginRegistrant
