+++ ios/.gitignore
@@ -34,3 +34,4 @@ Icon?
 .tags*
 
 /Flutter/Generated.xcconfig
+/Flutter/flutter_export_environment.sh
