+++ pubspec.yaml
@@ -1,11 +1,11 @@
-name: chirpsdk
-version: 1.0.0
+name: chirp_flutter
 description: Chirp enables your apps to send and receive information using sound.
+version: 1.0.0
 author: Asio Ltd <developers@chirp.io>
 homepage: https://github.com/chirp/chirp-flutter
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.1.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -17,5 +17,5 @@ dev_dependencies:
 
 flutter:
   plugin:
-    androidPackage: io.chirp.chirpsdk
-    pluginClass: ChirpsdkPlugin
+    androidPackage: com.example.chirp_flutter
+    pluginClass: ChirpFlutterPlugin
